general restructure of pixi
